feat: DAO, Application

+++ src/main/java/org/example/service/UserService.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public interface UserService {
     User createUser(String name, String email, Integer age, LocalDateTime created_at);
-    User readUser(int id);
-    void updateUser(int id, Integer age);
-    void deleteUser(int id);
+    User readUser(Long id);
+    void updateUser(Long id, Integer age);
+    void deleteUser(Long id);
 }
